debian/mediad.service: revived again
After various fixes about cgroups & Co., mediad seems to run as well
as systemd service. So revive this.
One problem is left, namely a message "Failed to attach ... to compat
systemd cgroup" because "mediad start" exits without forking if a
daemon was already running. I found a "ExitType=cgroup" might help
against this, but this isn't available yet in systemd-241, so
commented out (and live with the warning for now).